Use rapidfuzz for string matching

+++ include/utils/fuzzy_matcher.h
@@ -1,11 +1,10 @@
 #ifndef LAUNCHER_FUZZY_MATCHER_HPP
 #define LAUNCHER_FUZZY_MATCHER_HPP
 
-#include <cstdint>
 #include <string_view>
 
 namespace launcher::utils {
-    using fuzzy_match_score = int64_t;
+    using fuzzy_match_score = double;
 
     struct fuzzy_match_result {
         fuzzy_match_score score;
